# swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API (testing)",
    "description": "Does not work!"
  },
  "servers": [
    {
      "url": "https://veramo-jrtjw.ondigitalocean.app/",
      "description": "Public deployment",
      "variables": {}
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server",
      "variables": {}
    }
  ],
  "paths": {
    "/did/add": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "DID created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/didObject"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/didObject"
                }
              }
            }
          },
          "500": {
            "description": "Error creating DID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/addDid"
              }
            }
          }
        }
      }
    },
    "/did": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "List of all added DIDs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/didObject"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/didObject"
                }
              }
            }
          },
          "500": {
            "description": "Error retrieving DIDs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
    "/did/{didUrl}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "didUrl",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "DID URL",
            "example": "did:ethr:0xb09b66026ba5909a7cfe99b76875431d2b8d5190"
          }
        ],
        "responses": {
          "200": {
            "description": "Get DID by URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/didObject"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/didObject"
                }
              }
            }
          },
          "500": {
            "description": "Error retrieving DIDs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
    "/vc/issue": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "VC issued successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/credential"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/credential"
                }
              }
            }
          },
          "500": {
            "description": "Error issuing VC",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/issueVC"
              }
            }
          }
        }
      }
    },
    "/vc/verify": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "VC verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "boolean"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "boolean"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error verifying VC",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/credential"
              }
            }
          }
        }
      }
    },
    "/did/key/modify": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "Key added/removed/rotated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/didObject"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/didObject"
                }
              }
            }
          },
          "500": {
            "description": "Error modifying key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/modifyKey"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "string"
          }
        },
        "xml": {
          "name": "error"
        }
      },
      "addDid": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "example": "did:web:jakaskerjanc.github.io:alice"
          },
          "alias": {
            "type": "string",
            "example": "default"
          },
          "keyType": {
            "type": "string",
            "example": "Secp256k1"
          }
        },
        "required": [
          "provider",
          "alias",
          "keyType"
        ],
        "xml": {
          "name": "addDid"
        }
      },
      "didObject": {
        "type": "object",
        "properties": {
          "did": {
            "type": "string",
            "example": "did:web:jakaskerjanc.github.io:alice"
          },
          "alias": {
            "type": "string",
            "example": "default"
          },
          "provider": {
            "type": "string",
            "example": ""
          },
          "controllerKeyId": {
            "type": "string",
            "example": "did:web:jakaskerjanc.github.io:alice#key-1"
          },
          "keys": {
            "type": "array",
            "example": [],
            "items": {}
          },
          "services": {
            "type": "array",
            "example": [],
            "items": {}
          }
        },
        "required": [
          "did",
          "alias",
          "provider",
          "controllerKeyId",
          "keys",
          "services"
        ],
        "xml": {
          "name": "didObject"
        }
      },
      "issueVC": {
        "type": "object",
        "properties": {
          "didAlias": {
            "type": "string",
            "example": "default"
          },
          "credentialData": {
            "type": "object",
            "properties": {
              "key": {
                "type": "string",
                "example": "value"
              }
            }
          }
        },
        "required": [
          "didAlias",
          "credentialData"
        ],
        "xml": {
          "name": "issueVC"
        }
      },
      "credential": {
        "type": "object",
        "properties": {
          "credential": {
            "type": "object",
            "properties": {
              "@context": {
                "type": "array",
                "example": [
                  "https://www.w3.org/2018/credentials/v1"
                ],
                "items": {
                  "type": "string"
                }
              },
              "type": {
                "type": "array",
                "example": [
                  "VerifiableCredential"
                ],
                "items": {
                  "type": "string"
                }
              },
              "credentialSubject": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "example": "did:web:jakaskerjanc.github.io:alice"
                  },
                  "name": {
                    "type": "string",
                    "example": "Alice"
                  }
                }
              },
              "issuer": {
                "type": "string",
                "example": "did:web:jakaskerjanc.github.io:alice"
              },
              "issuanceDate": {
                "type": "string",
                "example": "2021-07-15T00:00:00Z"
              },
              "expirationDate": {
                "type": "string",
                "example": "2022-07-15T00:00:00Z"
              },
              "credentialStatus": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "example": "https://example.edu/status/24"
                  },
                  "type": {
                    "type": "string",
                    "example": "CredentialStatusList2017"
                  }
                }
              }
            }
          }
        },
        "required": [
          "credential"
        ],
        "xml": {
          "name": "credential"
        }
      },
      "modifyKey": {
        "type": "object",
        "properties": {
          "did": {
            "type": "string",
            "example": "did:ethr:0xb09b66026ba5909a7cfe99b76875431d2b8d5190"
          },
          "kidToRemove": {
            "type": "string",
            "example": "<public_key_in_hex_encoding>"
          },
          "action": {
            "type": "string",
            "example": "add"
          }
        },
        "required": [
          "did",
          "kidToRemove",
          "action"
        ],
        "xml": {
          "name": "modifyKey"
        }
      }
    }
  }
}